[osx] Updated xpackage target
gyp implements actions as pre-build steps in separate targets, so the first invocation of the versioning script was guaranteed to fail. Instead of a Run Script build phase, the version information can be added in xpackage, with the caveat that builds inside Xcode will report a bogus version number. In any case, we haven’t tagged any releases for OS X yet, so any version-related code would fail if we left it in.